PWM For STMF407VGT
Adding in the pwm code but will update to reflect the blue pill stm32F1 mcu we will be using. PWM set at 50Hz for no though is easily changed to whatever desired frequency, pwm file contains the functions necessary to set pwm for a port, pwm_example is just an example of how pwm works with HAL on stm board.